revise: updates filters to accept valid JSON values
Summary
=======

- Changes filter parameters to accept only valid JSON objects. This change was
  introduced to allow for more accurate filtering. For example, `null` can now
  be filtered on. This also means that strings are required to be surrounded in
  quotes.
- Explicitly outlines how each field should be filtered. Note that some
  fields are implemented in a non-standard way based on the contemplated use
  cases for them. As such, implementers should carefully read each field to
  understand how the fields should be filtered.
- The `identifier` and `diagnosis` filters were added to `Sample`. These
  appear to have just been missed when adding these fields. All harmonized
  metadata fields should now have a filter parameter.
- Changes plural filter parameter names (e.g., `identifiers`) to
  singular filter parameter names (e.g., `identifier`).

Other Notes
===========

- Adds a checklist item to ensure filtering works properly.

+++ packages/ccdi-models/src/file/metadata.rs
@@ -243,7 +243,7 @@ impl Metadata {
                 None,
             )),
             size: Some(field::unowned::file::Size::new(
-                cde::v1::file::Size::new(rand::thread_rng().gen_range(usize::MIN..=usize::MAX)),
+                cde::v1::file::Size::new(rand::thread_rng().gen_range(u64::MIN..=u64::MAX)),
                 None,
                 None,
                 None,
